CORE-17397: Turn off tracing logs in kubernetes (#4733)
This PR turns off logging trace spans in both the console and helm chart log4j configuration. These trace spans are intended to go into a log file and they should be switched off by default.

It also changes the existing configuration to be very specific about which class we want to send to the `corda-trace-spans` log file. So we don't sweep logs related to the operation of tracing into that file.