CORE-17391 Set registration state command metadata (#4730)
With this change, after each successful registration command processing, if the state has not been nullified, the command which was processed is added to a tracking list along with an index to allow future stages to determine which commands have previously been processed. This allows processing to be skipped in cases that it has already been completed. In this PR, the process member verification response processor is updated so that it does not process again if it has already done so. This protects against replays at the P2P level.