CORE-17357: Create a ScheduledTaskHandler for processing scheduled tasks in the flow mapper (#4723)
Adds a class for handling scheduled task triggers to the flow mapper.

The flow mapper cleanup logic requires integrating with the new scheduler to function once the new state management functionality is introduced. To do this, there needs to be a component in the flow mapper responsible for listening to scheduled tasks and fanning out the corresponding IDs to the set of mapper processors.