CORE-17358: Add a class for cleaning up mapper states (#4743)
Adds a handler for processing mapper cleanup events, to be used from a mapper cleanup processor.

The flow mapper cleanup mechanism needs to be replaced, as the current mechanism will no longer function when the mapper is migrated to use the new state manager and event mediator APIs. Completing this requires a class responsible for picking up ExecuteCleanup events and removing the corresponding states.

For now, this class is not plugged in anywhere. No assumptions have been made about where these events are published, although it is assumed they will be sent to a new topic to keep the cleanup events off the main flow mapper event topic.