CORE-9031: vNode upgrade - forceUpgrade parameter to override operation in progress (#4734)
This adds a "forceUpgrade" query parameter to the vNode upgrade endpoint which allows the upgrade to bypass validating that there is no OperationInProgress on the vNode.

This is being added to allow fixing a vNode that has failed during an operation and been left in an invalid state with a non null OperationInProgress field. By allowing an upgrade to be forced, this will provide a way to reset the OperationInProgress in this scenario.