CORE-17352 - Prevent notary vnodes from using ledger key (#4751)
Adjustments to ensure a notary vnode cannot register with a ledger key in a dynamic network. In static networks, keys are not specified as part of registration, they are auto-generated & registered from the platform, so the code has been adjusted to not register ledger keys for notaries for a consistent behaviour. The CLI plugin was also adjusted to not specify a ledger key, when onboarding a notary.