CORE-14025 Check if transaction present in the database before requesting in backchain resolution (#4592)
Adding a new request to the UTXO ledger layer to fetch transaction that are not invalid. This means they have verified or unverified status.

This is needed when doing a back-chain resolution. We don't want to fetch transactions that are already present in our database. We just want to fetch their IDs though, no need to deserialise the whole object.

The back-chain receiver flow will now check if any of the transactions we want to fetch are already in the database. If yes, then we'll just not request those transactions.