CORE-17548 Fix Backchain state resolution error
We recently added a call to `findSignedLedgerTransactionWithStatus`
which involves resolving input and reference states when performing
backchain resolution. However, there is a possibility that a transaction
which we're checking the existence of in the database is UNVERIFIED and
has not yet resolved its dependencies will error when
`findSignedLedgerTransactionWithStatus` is called.

 This occurs because the dependent transactions haven't been resolved
 and the state resolution in `findSignedLedgerTransactionWithStatus`
 fails.

 To fix this, we instead call `findSignedTransactionWithStatus` which
 doesn't execute state resolution.