CORE-17596 Adding service endpoints in MessagingConfig with support for CombinedWorker (#4823)
Previously, a `ClusterIP` service was added to each worker, acting as a load balancer, and the endpoints were being passed in as startup arguments to the `FlowWorker`. Now we're adding these to the `BootConfig` and copying them across to the `MessagingConfig` which is used to create the subscription which backs the flow engine. 

This PR also defaults these endpoints to `localhost` in the `CombinedWorker`.