CORE-17509 Fix MGM platform version after platform upgrade (#4822)
Fixes an issue where the platform version in the exported group policy would continue to be the older version after performing a platform upgrade. While exporting the group policy, the platform version is now retrieved from PlatformInfoProvider instead of the MGM's MemberInfo.
This change also republishes the MGM's MemberInfo before re-registering a member (after the virtual node has been upgraded), if the Group Policy was changed.