CORE-15552 Add `findUnconsumedStatesByExactType` paging API
Reuse the vault named query logic to create a
`findUnconsumedStatesByExactType` API that supports paging without too
much extra work.

`DefaultVaultNamedQueryFactoryImpl` registers a query for finding states
by their exact type.

This factory is then manually registered in
`VaultNamedQueryFactoryProvider`.

We could potentially add a better lookup mechanism in the future, but
there is unlikely to be additional `VaultNamedQueryFactory`s that we
register in the future.