CORE-16323: Fix State Manager Optimistic Locking (#4846)
Fix update and delete SQL statements within the State Manager
implementation so only the states with the specified key and
version are updated. This prevents race conditions while executing
the optimistic locking and also reduces by half the amount of SQL
statements that needs be executed for every update and delete.