CORE-16234: For persistence api add endpoint and change naming of existing endpoint (#4774)
For the RPC Epic the Persistence api will require two endpoints. One endpoint is already merged which relates to ledger-related persistence. This PR will be making a suggested change to the naming of that endpoint for clarity, and will also add the second endpoint required.

When this work is merged we will have the endpoints:
/ledger - for ledger persistence
/persistence - for entity persistence