CORE-16419 - template, schema and values yaml changes to support isolated state manager db (#4763)
These changes enable an isolated state-manager database to be automatically deployed via the Corda helm chart when db bootstrapping is enabled. 

A state-manager database can be bootstrapped by providing the `bootstrap.db.stateManager` postgres credentials overrides, and a `stateManager.db.host` override, along with credentials in `stateManager.db`. 

If no overrides are provided, the state-manager will be bootstrapped in a schema in the cluster database with cluster credentials. Template functions in _helpers.tpl will fall back to cluster JDBC URL and credentials.