CORE-17625 MGM lookups return active platform version in its own MemberInfo (#4855)
After performing a platform upgrade, an MGM’s view of its own MemberInfo is not refreshed, so any lookups performed by such an MGM would continue to show it to be on the older version. This change updates the MembershipGroupReader to decorate the returned MGM MemberInfo (when requested by the MGM) with the active platform version from PlatformInfoProvider. This will be removed on completion of CORE-17811, once support for MGM re-registration is available.