CORE-17885: Use explicit conversion of session timeout to long (#4927)
Adjusts the session timeout processor to use an explicit conversion to long.

When going through a cycle of writing metadata to the state store and reading it out again, the original type of any numeric values is lost. This causes a problem as casts between numerics are not allowed. To address this, the underlying numeric value in the metadata is converted to `Number`, and then explicitly converted to a long.

This additionally introduces a more thorough unit test for the flow event mediator factory, which ensures that routes are correct for different event types.