CORE-16983 Prerequistites for Serialize Merkle Proof part of CORE-16983 ticket - test signature proof (#4962)
Test signature proof generation by a notary, in the following PR we do add serializers and deserializers for MerkleTree Proof and some other things in order to send them via RPC/or Corda sessions.
This is the preliminary PR to add ability for UtxoLedgerTest to sign a batch transaction in order to generate Merkle Tree proof. The existing UtxoLedgerTest had hardcoded signatures without calling singing service (or a mock), and there was never a proof produced (as not needed so far). I intentionally created a new class UtxoLedgerWithBatchSignerTest inheriting from UtxoLedgerTest just not to clutter the original one, but it could be added to UtxoLedgerTest directly.
This feature as part of full serialization/deserialization functionality will be PR against 5.2.