CORE-17389: Process flow timeout events and clean up checkpoints (#4951)
Problem description
The flow engine currently responds to scheduled task triggers to produce flow timeout events when a session has timed out. These need to be processed to remove the timed out checkpoint and send any cleanup events (session errors and mapper cleanup events) to the rest of the system. The flow status also needs to be updated.

Solution
Add a processor for the flow timeout events.

This PR also includes a checkpoint cleanup handler class, which generates the relevant events for when a flow is terminated with an error. This can be reused in the FlowEventExceptionProcessor, although this has not been integrated in this PR. The objective here is to encapsulate this cleanup logic. At present, there is a bug in the flow engine that means not all of this logic is present everywhere it is needed.