DA5-8 Prerequisites for Serialize Merkle Proof - test base (#5000)
Test signature proof generation by a notary. Add ability for UtxoLedgerTest to sign a batch transaction in order to generate Merkle Tree Proof. The existing UtxoLedgerTest had hardcoded signatures without calling signing service (or a mock), and no test has created a proof so far (as it was not needed).