DA5-8 MerkleTree Proof serializers and deserializers (#5006)
Add serializers and deserializers for MerkleProof and other related classes. New 'InteropJsonMarshallingServiceImpl' implementing new InteropJsonMarshallingService uses those serializes. The serializer needs MerkleProofImpl which are getting via another service MerkleProofProvider / MerkleProofProviderImplementation.
The new serializers are needed to pass signatures as arguments of an RPC Flow invocation and as a response from RPC Flow. For now the new service is created separately to existing one existing JsonMarshallingServiceImpl - the new service has the same code as existing one with extra MerkleProof etc, however the new service is only needed inside a flow. We could merge both services in one in the future.