DA5-8 Unit tests for MerkleProof JSON serialization (#5025)
Added test for DigitalSignatureAndMetadataSerialisationModule (renamed now from ProofOfActionSerialisationModule) which was created in aa29e22.
The test requires a transaction signature with a MerkleProof. This type of signature can be produced by the base test class UtxoLedgerTest available for use in components/ledger/ledger-utxo-flow module where the test is placed.
The test requires a service implementation of MerkleProofProvider which is injected into the SUT, the implementation is in libs/crypto/merkle-impl hence with Gradle dependency testImplementation project(':libs:crypto:merkle-impl').
The actual test the class DigitalSignatureAndMetadataSerialisationModule is from libs/application/application-impl, so the test is not placed in the same module as the class.
There are 2 reasons: application-impl module has no test dependencies on either testing/ledger/ledger-utxo-base-test or libs/crypto/merkle-impl, while adding it to components/ledger/ledger-utxo-flow requires addtion of the second latter dependency only.
The second reason is that if the test is run from withing a module other than components/ledger then it throws an exception that Sandbox cannot be mocked.
The other placing of the test could me moved later, however putting it inside components/ledger/ledger-utxo-flow looked as reasonable compromise.