CORE-17966: Implement checkpoint cleanup handler when processing FlowFatalException (#5044)
Change:
This PR will implement checkpoint cleanup handler when processing FlowFatalException.

Reasoning:
In order to address CORE-17966: When exceptions such as FlowFatalException are thrown in the Flow Engine the exception handlers cleanup the checkpoints and schedule cleanup of the session states. However they currently do not cleanup the StartFlow Flow Mapper state for RPC started flows the checkpoint cleanup handler will be integrated in a series of small PRs.