CORE-17194 Adding prefix to consumer group names (#5055)
This change adds a prefix to consumer group IDs to support multi-tenancy (>1 Corda cluster within a single Kafka cluster), in the same way that we currently do for topic names. Without this, consumers in different Corda deployments were being considered to be part of the same consumer group; this causes unnecessary re-balances, and was logging authz exceptions on partition assignment.