CORE-17917 Adding handling to MessageBusClient for async kafka errors (#5107)
A bug was observed whereby Kafka producer errors were not being correctly reported back to the MultiSourceEventMediator. This is because our Kafka messages are essentially 'fire-and-forget'; when something goes wrong, Kafka will exceptionally close a future in the background, but we never attempted to capture or track this future. This PR adds a callback which will log this exception clearly for the user.