CORE-12165: merge key rotation feature work (#5126)
Take new API update with additional Avro, kafka and config for crypto key rotation.

Add new RestResource called KeyRotationRestResource which is automatically injected into the REST procesosr and allows crypto unmanaged wrapping key rotation to be started.

Implement unmanaged key rotation, for now in a relatively simple way where we take work requests on Kafka then process them to completion. This won't necessarily be the best way for the database at scale, and we have a design for spreading the work out over a nested Kafka topic, but let's get something in that is simple and works first.

Add alias to WrappingKeyInfo. It's not clear why this wasn't here before. WrappingKeyInfo is a data transfer object (DTO) and it seems reasonable for it to hide DB implementation detals such as the synthetic primary ID field. Alias is also part of a compound key with generation, and both alias and generation are user-visible features of wrapping keys; for instance a user could specify wrapping key alias when creating a signing key pair.

Making WrappingKeyEntity key_material updatable since that's what happens during key rotation.

Add state manager arguments to the crypto worker, so that later the crypto worker can use the state manager for status notifications.