CORE-18492: register notary with backchain required field (#5167)
To be able to implement a contract verifying notary which allows skipping backchain and does contract verification within notary, we need to have a backchainRequired flag and register the notary with this flag. This PR added the flag in membership registration and cascaded the field with respect to it.

The backchainRequired field is true in default, considering a non-validating notary to be the default notary, which requires backchain resolution.