CORE-18030: Batch up state creation in the state manager (#5180)
Batch state creation in the state manager to reduce database queries.

Previously if a group of states was requested to be created the state manager would create each in an individual transaction. With this change, a single query is used to create any states that do not already exist.

The state manager API has also been modified to remove the `Exception` from the return type of create. This brings it in line with other state manager APIs.