CORE-18579: Provide types as class parameters (#5187)
The processor implementation in workers have requestClass and responseClass included in their constructors. This was changed in the Entity Persistence processor to be provided as class parameters. This PR is to change other workers - Persistence (of type Ledger), Verification and Crypto - for consistency in the pattern.
Note - Uniqueness was changed in a uniqueness refactoring PR that was already in-flight #5184