CORE-18724 Add registration logger and update logging during registration to aid tracking registration progress (#5217)
From time to time we need to investigate failed registrations from application logs. Our logging through registrations can be a bit inconsistent and when we have a high volume of registrations (for example during large network or resiliency testing) it can be hard to follow the progress of a single registration. Some logs have registration ID, member holding ID, member name, MGM holding ID, group ID, some combination of those, or none of those.
This PR adds a registration logger which is a wrapper around the log4j logger and includes registration metadata as part of logging lines in a standard way so we can more easily piece together at which stage a registration failed. It also adds logging to some areas we were lacking previously.