Add serializers and deserializers for MerkleProof and other related classes.'JsonMarshallingServiceImpl' will use the serializes and deserializers. The new serializers are needed to pass signatures as arguments of an RPC Flow invocation, and it's response. Some of new serializers need MerkleProofImpl which is provided via another new service MerkleProofProvider.
Add testing of the new DigitalSignatureAndMetadataSerialisationModule module. The test requires a transaction signature with a MerkleProof. This type of signature can be produced by the base test class UtxoLedgerTest available for use in components/ledger/ledger-utxo-flow module where the test is placed. The test requires a service implementation of MerkleProofProvider which is injected into the SUT, the implementation is in libs/crypto/merkle-impl hence with Gradle dependency testImplementation project(':libs:crypto:merkle-impl'). The actual test the class DigitalSignatureAndMetadataSerialisationModule is from libs/application/application-impl, so the test is not placed in the same module as the class. There are 2 reasons: application-impl module has no test dependencies on either testing/ledger/ledger-utxo-base-test or libs/crypto/merkle-impl, while adding it to components/ledger/ledger-utxo-flow requires addition of the second latter dependency only.
The second reason is that if the test is run from withing a module other than components/ledger then it throws an exception that Sandbox cannot be mocked. Placing the test it inside components/ledger/ledger-utxo-flow looked as reasonable compromise.