CORE-17596 Setting default service endpoints in combined worker (#5237)
The `FlowWorker` and `CombinedWorker` need service endpoints to be set for communication with external workers over HTTP/RPC. For the `FlowWorker`, these are provided through helm. For the `CombinedWorker`, this endpoint should be localhost; previously this was provided as a command line arg, but now it's defaulted within `CombinedWorker.kt`.