CORE-18691: More comments, render Merkle trees for test cases, add calculateLeveledHash code (#5226)
Increase the detail on the comments in MerkleProofImpl and MerkleTreeImpl, to make the code easier to maintain.

Improve test coverage on Merkle trees by showing the content and root hashes, including draw out the tree structure.
Next step is to add rendering of Merkle proofs so that I can assert the whole structure of the proof by matching one string, at which point this unicode rendering will be more useful.

For example, we now have an assertion that our test input data, in the form of sequential 32 bit ints, hashes like this:

              a868a19c7 ┳┳┳ 00:00:00:00
                        ┃┃┗ 00:00:00:01
                        ┃┗┳ 00:00:00:02
                        ┃ ┗ 00:00:00:03
                        ┗┳┳ 00:00:00:04
                         ┃┗ 00:00:00:05
                         ┗┳ 00:00:00:06
                          ┗ 00:00:00:07
That's actually all the infromation we reveal in the Merkle Tree implementation right now. The hashes of non-root nodes is not accessible, nor can you tell the tree structure.  The unicode illustration of the structure within is actually a detail of our Merkle tree implementation - it's always a left binary balanced. Since we don't expose the tree structure, when it comes to talking about Merkle proofs it's hard to know where the proof hashes go, unless you reimplement the proof code.

Previously we don't actually test the root hash of trees, so we could for instance accidentally change the tree balancing which would change the root hash. Since we now include the root hash and check the pretty-printed Merkle tree hashes we are improving test coverage. There's scope for more coverage and we can add that later. I've also added some assertions I did about specific values within the trees/proofs.

Later, we could extend the text representation with the unicode to proofs, where we can show intermediate hashes filled in by the proofs.

Since MerkleTreeImpl and MerkleProofImpl do not reveal the structure of the tree, we introduce calculateedLeveledHashes, which takes an audit proof and works out the levels and indices within that level of each hash required to fill in gaps when verifying a proof (i.e. the proof hashes). This code has a small amount of test coverage for now, including some negative cases. We run the code over the Merkle proofs we come up with during other tests, and check a little bit of the level information. For now this duplicates code, and the motivation to improve test coverage it to make it safe to move ahead and reduce duplication once we have solid test coverage, in future work. The hope is that this extra information somehow helps serialise proofs and re-use proof elements.

Co-authored by Dan Newton, i.e. based on his POC.