CORE-18320 Support for a flow specific timeout on Flow Message API (#5247)
Support for a flow specific timeout on Flow Message API. Client can specify a flow session specific timeout on Flow Message API. If set, it will be passed to the initiated party to also use it. If the client does not provide a timeout value then the value from Corda Configuration is used.

Session timeout is in SessionState stored in field sessionProperties. With this PR requireClose is also stored in sessionProperties instead of in separate field.