CORE-18961 Thread safe SandboxServiceImpl
When the sandbox cache is stressed a bit it tries to load and unload OSGi bundles concurrently, because the close() block is executed in a thread which triggers a purge, whilst another thread can be creating a new sandbox at the same time. There were various maps and sets in this service being manipulated by multiple threads with no locking.