CORE-18961 Thread safe SandboxServiceImpl (#5325)
When the sandbox cache is stressed it tries to load and unload OSGi bundles concurrently. This is because the close() block is executed in whatever thread happens to trigger a purge, whilst another thread can be creating a new sandbox at the same time. Both of these operations interact with SandboxServiceImpl which is our abstraction for OSGi bundle interaction. This service contains a number of maps and sets which would be manipulated concurrently at this point, and this is not thread safe.

I could reproduce this consistently by reducing the sandbox cache size to 1. With the introduction of a small amount of extra logging I could see the concurrent creation/closing of sandboxes from multiple threads followed immediately by a "bad wiring" exception from OSGi as it then tried to process the flow message which resulted in a new sandbox in a bundle which was incorrectly configured.

Making the SandboxServiceImpl thread safe resulted in the "bad wiring" errors going away, whilst many instances of unloading of bundles could still be observed in the log (this was whilst sandbox cache size was still forced to 1).