CORE-17886: Protect Javalin server creation with `serverLock` (#5327)
When REST worker starts in `RestServerInternal` class it makes `addStaticFiles("/META-INF/resources/", ...)` call in the context of SwaggerUIClassloader.
This call is happening at `RestServerInternal` instantiation time and not protected by the lock. This call not just adds static files to the webserver, but also executes `JavalinLogger.addDelayed` call to log a line in the log at a later point.

When another Javalin instance starts for HTTP RPC processing (e.g. in Crypto) and if it obtains a lock ahead of `RestServerInternal`, it invokes `logAllDelayed` method. But, unfortunately, at this point in time SwaggerUIClassloader is no longer available and "/META-INF/resources/" can no longer be retrieved, hence logging fails. As an unfortunate side effect, this prevents Javalin HTTP RPC server from starting rendering Combined Worker in the unhealthy state.

In essence, before the change we had a race condition between Javalin server creation and start-up. After the change Javalin server creation and start-up are all happening under global lock protection.