CORE-18691 Sandbox Logging (#5349)
Adding more logging to sandbox creation and also bundle installation and uninstallation so we can ensure the correct sandbox is being used at the point of failure. We already have logging for sandbox eviction and closing and so this extra logging adds some symmetry to that.

Also removed a function we only ever called in tests. This reduces some complexity in the SandboxCreationService which has no need for "non started" bundles.