CORE-18706 - Fixed external messaging after the mediator pattern was introduced (#5348)
The mediator pattern introduced a translation from `Record< * , *>` to `MediatorMessage<*>`. There are two key points to keep in mind. The field `topic` in the `Record<*,*>` is dropped and the `payload` field in the `MediatorMessage<*>` contains the message to be sent when the `external messaging` feature is used. In order to fix the issue, the value in the `topic` field is now added to the `properties` field in the `MediatorMessage<*>` which is later used by `FlowEventMediatorFactoryImpl`.