CORE-18978 Remove Kafka schema validation on reconciliation (#5368)
It is unnecessary to validate the schema from Kafka (this is old Config which is about to be reconciled with the new schema if it is incorrect anyway), and doing so (unexpectedly) adds defaults even though we ask it not to. This breaks the reconciliation check on upgrades as the db worker believes new fields are already published to Kafka when they are not.

A separate JIRA ticket is raised to cover the problem with validate, it's not fixed here because given we don't use it any more it's now unrelated to this change, and also quite low priority to investigate an unused method.