CORE-19168, CORE-19176: show intermediate hashes in merkle tree and proof representations (#5362)
Extend the text representation of Merkle trees to include intermediate hashes. This makes the way the trees work more transparent in the unit tests.

SInce the intermediate hashes are private this involves moving the tree rendering code to be available from a toString() method. I also extended the tree rendering code to allow labels on all nodes as well as just leaves, and I took the time to document the rendering routine better.

Use ints rather than longs for the trivial hashes so that the test cases look nicer. The way the trivial hashes are done the least significant bits are the distinctive ones so we can't just truncate in renders.

Rendering a Merkle proof requires using the root calculation algorithm with additional bookkeeping, so we add a callback to that code.