CORE-18535: Fail processing if the retry limit is exceeded and mark state with failed metadata key (#5396)
Fail processing of a set of events if the HTTP RPC retry limit is exceeded.

Previously the event mediator would get into an infinite loop if the synchronous calls failed consistently. This PR changes this behaviour to instead fail processing and mark the state as having failed processing. The client may query for states that have failed processing in order to handle this failure.