CORE-18534 remaining external event processors (#5400)
This is the 2nd PR to update external event processors to throw CordaHTTPServerTransientExceptions for transient errors. It includes:

- LedgerPersistenceRequestProcessor
- TokenSelectionSyncRPCProcessor
- VerificationRequestProcessor

Problem description:
All flow engine external events go through the RPCClient in the messaging library. This has a built-in retry mechanism to retry events for some period of time. If an external event hits a transient error in the external event handler, the flow engine should rely on this retry handling rather than implementing its own.

To do this, the external event handlers must always return a 503 retry code if a transient error is encountered. The messaging library handles this, the processor throws the exception mentioned above.