CORE-19016 store async outputs (#5426)
The list of asynchronous output records associated with each consumer record polled and processed needs to be saved to the state manager.

The mediator already groups asynchronous outputs together into a collection. This collection is altered to store them keyed by input event.
A new service is provided to handle mediator replay logic

Output events are stored in MediatorState. Events should be keyed by consumer record input hash to allow for replays to be detected.