CORE-18518 - Baseline performance testing for P2P (#5344)
For performance testing purposes we needed to modify the scripts to be able to generate two-way load scenarios. We would also like to test scenarios where multiple members are present on the network, so the onboarding script was modified accordingly. These members should be able to communicate with everyone from the other cluster, the Sender code needed some slight adjustments to achieve this.