CORE-18930: make transaction ID deterministic (#5443)
The privacy salt currently gets a unique ID for transactions that is random. We need it to be unique but also deterministic. The random aspect will be replaced with a concatenation of FlowID+suspendCount+counter which are obtained from the checkpoint. To get this data, FlowFiberService will be used. A service will be created that uses FlowFiberService to give our privacy salt, so this can be injected in several places. Testing should show that transactions with the same flowID and suspendCount and counter give the same ID (deterministic) even if created multiple times.